+++ rest_framework_reactive/routing.py
@@ -1,8 +1,21 @@
-from .consumers import ObserversConsumer
+from django.conf.urls import url
 
+from channels.routing import ChannelNameRouter, ProtocolTypeRouter, URLRouter
 
-default_channel_routing = [
-    # To change the prefix, you can import ObserversConsumer in your custom
-    # Channels routing definitions instead of using these defaults.
-    ObserversConsumer.as_route(path=r'^/ws/(?P<subscriber_id>.+)$'),
-]
+from .consumers import ClientConsumer, PollObserversConsumer, WorkerConsumer
+from .protocol import CHANNEL_POLL_OBSERVER, CHANNEL_WORKER_NOTIFY
+
+application = ProtocolTypeRouter({
+    # Client-facing consumers.
+    'websocket': URLRouter([
+        # To change the prefix, you can import ClientConsumer in your custom
+        # Channels routing definitions instead of using these defaults.
+        url(r'^ws/(?P<subscriber_id>.+)$', ClientConsumer),
+    ]),
+
+    # Background worker consumers.
+    'channel': ChannelNameRouter({
+        CHANNEL_POLL_OBSERVER: PollObserversConsumer,
+        CHANNEL_WORKER_NOTIFY: WorkerConsumer,
+    })
+})